Ponyo Feels - Containerized App w: Machine Learning Client/requirements.txt
flask
python-dotenv
pymongo
hume
pytest
coverage
pylint
chardet
coverage
hume[stream]